refactor: composite effects architecture
