+++ src/effects/sleep.rs
@@ -12,8 +12,8 @@ impl Effect for Sleep {
         let len = pixels.len();
         let total_frames = (self.duration * 60) as usize;
         let start_color = pixels[0];
-        let effect_start_brightness = self.start_brightness;
-        let effect_target_color = self.target_color;
+        let start_brightness = self.start_brightness;
+        let target_color = self.target_color;
 
         let mut frame_count = 0;
         Box::new(std::iter::from_fn(move || {
@@ -22,9 +22,9 @@ impl Effect for Sleep {
             }
 
             let progress = frame_count as f32 / total_frames as f32;
-            let current_brightness = effect_start_brightness as f32 * (1.0 - progress);
+            let current_brightness = start_brightness as f32 * (1.0 - progress);
 
-            let interpolated_color = Rgb::lerp(&start_color, effect_target_color, progress);
+            let interpolated_color = Rgb::lerp(&start_color, target_color, progress);
             let final_color = interpolated_color.with_brightness(current_brightness.round() as u8);
 
             frame_count += 1;
