+++ src/effects/wake.rs
@@ -11,8 +11,8 @@ impl Effect for Wake {
     fn frames(&self, pixels: &[Rgb]) -> Box<dyn Iterator<Item = Vec<Rgb>> + Send + 'static> {
         let len = pixels.len();
         let total_frames = (self.duration * 60) as usize;
-        let initial_color = self.start_color;
-        let effect_end_brightness = self.end_brightness;
+        let start_color = self.start_color;
+        let end_brightness = self.end_brightness;
 
         let mut frame_count = 0;
         Box::new(std::iter::from_fn(move || {
@@ -21,12 +21,8 @@ impl Effect for Wake {
             }
 
             let progress = frame_count as f32 / total_frames as f32;
-            let current_brightness = effect_end_brightness as f32 * progress;
-
-            let target_color_at_full_brightness = initial_color.with_brightness(255);
-            let interpolated_color =
-                Rgb::lerp(&initial_color, target_color_at_full_brightness, progress);
-            let final_color = interpolated_color.with_brightness(current_brightness.round() as u8);
+            let current_brightness = end_brightness as f32 * progress;
+            let final_color = start_color.with_brightness(current_brightness.round() as u8);
 
             frame_count += 1;
             Some(vec![final_color; len])
